Split secret providers
