+++ k8t/secret_providers/__init__.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+# ISC License
+#
+# Copyright 2019 FL Fintech E GmbH
+#
+# Permission to use, copy, modify, and/or distribute this software for any purpose with or without fee is hereby granted, provided that the above copyright notice and this permission notice appear in all copies.
+#
+# THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
+import importlib
+import logging
+
+from k8t import config
+
+LOGGER = logging.getLogger(__name__)
+
+def get_secret(key: str, length: int = None) -> str:
+    provider_name = config.CONFIG.get("secrets", {}).get("provider")
+    if not provider_name:
+        raise RuntimeError("Secrets provider not configured.")
+
+    provider_name = str(provider_name).lower()
+    try:
+        provider = importlib.import_module(f".{provider_name}_provider", package=__name__)
+    except ModuleNotFoundError:
+        raise NotImplementedError(f"secret provider {provider_name} does not exist.")
+
+    return provider.get_secret(key, length)
