+++ tests/secret_providers/random_provider.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+# ISC License
+#
+# Copyright 2019 FL Fintech E GmbH
+#
+# Permission to use, copy, modify, and/or distribute this software for any purpose with or without fee is hereby granted, provided that the above copyright notice and this permission notice appear in all copies.
+#
+# THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
+import pytest
+from k8t.secret_providers.random_provider import get_secret
+
+
+def test_get_secret():
+    assert get_secret("/foobar") == get_secret("/foobar")
+    assert get_secret("/foobar") != get_secret("/foobaz")
+
+    assert len(get_secret("/foo", 12)) == 12
+    with pytest.raises(AssertionError, match=r"Secret '/foo' did not have expected length of 3"):
+        get_secret("/foo", 3)
+
+# vim: fenc=utf-8:ts=4:sw=4:expandtab
